fix: refactored defi opportunity functions
